working well so far might need to edit hook abit
