+++ modules/dateCode.js
@@ -44,3 +44,10 @@ export function filterTimeboxesBasedOnWeekRange(timeboxes, selectedDate) {
     timeboxes = timeboxes.slice(0, indexOfEndOfRange+1);
     return timeboxes;
 }
+
+export function differenceInDates(date1, date2) {
+    let firstDate = dayjs(date1).hour(0).minute(0).second(0).millisecond(0);
+    let secondDate = dayjs(date2).hour(0).minute(0).second(0).millisecond(0);
+    console.log(firstDate.diff(secondDate, 'day'), firstDate, secondDate);
+    return firstDate.diff(secondDate, 'day');
+}
